+++ app/studio/inspector/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import IntelligenceInspector from "@/components/studio/tools/IntelligenceInspector";
+
+export default function StudioInspectorPage() {
+  return <IntelligenceInspector />;
+}