+++ lib/studio/tools/index.ts
@@ -1,0 +1,4 @@
+export * from "./demoDirector";
+export * from "./learnerSimulator";
+export * from "./intelligenceInspector";
+export * from "./eventMonitor";